extract recent routes
